move lua libs to /logic/scripting/lua/libs
