+++ src/logic/scripting/lua/libs/libcamera.cpp
