+++ src/logic/scripting/lua/libs/libconsole.cpp
