+++ src/logic/scripting/lua/libs/libcore.cpp
