+++ src/logic/scripting/lua/libs/libquat.cpp
