+++ src/logic/scripting/lua/lua_extensions.cpp
@@ -1,6 +1,6 @@
 #include <iostream>
 
-#include "api_lua.hpp"
+#include "libs/api_lua.hpp"
 #include "debug/Logger.hpp"
 
 static debug::Logger logger("lua-debug");
