+++ src/logic/scripting/lua/lua_overrides.cpp
@@ -1,6 +1,6 @@
 #include <iostream>
 
-#include "api_lua.hpp"
+#include "libs/api_lua.hpp"
 
 /// @brief Modified version of luaB_print from lbaselib.c
 int l_print(lua::State* L) {
